pyproject.toml
[project]
name = "kamateratoolbox"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "jinja2==3.1.2",
    "paramiko==2.7.1",
    "pexpect==4.8.0",
    "pytest==5.4.3",
    "pytest-xdist==1.32.0",
    "python-dotenv>=1.2.1",
    "requests==2.32.3",
    "ruamel-yaml==0.16.10",
    "urllib3==2.2.3",
    "watchdog==2.1.9",
]
